fix sorting, refactor table generation
