+++ packages/syft/src/syft/service/code_history/code_history.py
@@ -10,9 +10,10 @@ from ...service.user.user_roles import ServiceRole
 from ...types.syft_object import SYFT_OBJECT_VERSION_2
 from ...types.syft_object import SyftObject
 from ...types.syft_object import SyftVerifyKey
-from ...types.syft_object import _get_table_data
 from ...types.uid import UID
 from ...util.notebook_ui.components.table import create_table_template
+from ...util.table import _create_table_rows
+from ...util.table import prepare_table_data
 from ..code.user_code import UserCode
 from ..response import SyftError
 
@@ -55,7 +56,8 @@ class CodeHistoryView(SyftObject):
         return {"Number of versions": len(self.user_code_history)}
 
     def _repr_html_(self) -> str:
-        rows = _get_table_data(self.user_code_history, True)
+        # TODO techdebt: make _coll_repr_ instead of _repr_html_
+        rows, _ = prepare_table_data(self.user_code_history)
         for i, r in enumerate(rows):
             r["Version"] = f"v{i}"
             raw_code = self.user_code_history[i].raw_code
