+++ packages/syft/src/syft/types/syft_object.py
@@ -12,7 +12,6 @@ from collections.abc import Set
 from hashlib import sha256
 import inspect
 from inspect import Signature
-import re
 import types
 from types import NoneType
 from types import UnionType
@@ -26,8 +25,6 @@ from typing import get_args
 from typing import get_origin
 
 # third party
-from loguru import logger
-import pandas as pd
 import pydantic
 from pydantic import ConfigDict
 from pydantic import EmailStr
@@ -44,8 +41,7 @@ from ..serde.recursive_primitives import recursive_serde_register_type
 from ..serde.serialize import _serialize as serialize
 from ..util.autoreload import autoreload_enabled
 from ..util.markdown import as_markdown_python_code
-from ..util.notebook_ui.components.table import TABLE_INDEX_KEY
-from ..util.notebook_ui.components.table import create_table_template
+from ..util.table import list_dict_repr_html
 from ..util.util import aggressive_set_attr
 from ..util.util import full_name_with_qualname
 from ..util.util import get_qualname_for
@@ -764,248 +760,6 @@ def short_uid(uid: UID | None) -> str | None:
         return uid
     else:
         return str(uid)[:6] + "..."
-
-
-def _sort_table_rows(rows: list[dict[str, Any]], sort_key: str) -> list[dict[str, Any]]:
-    # relative
-    from .datetime import DateTime
-    from .datetime import str_is_datetime
-
-    if sort_key not in rows[0]:
-        return rows
-
-    sort_values = [row[sort_key] for row in rows]
-    if isinstance(sort_values[0], str) and str_is_datetime(sort_values[0]):
-        sort_values = [DateTime.from_str(v) for v in sort_values]
-
-    reverse_sort = False
-    if isinstance(sort_values[0], DateTime):
-        sort_values = [d.utc_timestamp for d in sort_values]
-        reverse_sort = True
-
-    rows_sorted = [
-        row
-        for _, row in sorted(
-            zip(sort_values, rows),
-            reverse=reverse_sort,
-            key=lambda pair: pair[0],
-        )
-    ]
-
-    return rows_sorted
-
-
-def _get_table_data(
-    _self: Mapping | Iterable,
-    is_homogenous: bool,
-    extra_fields: list | None = None,
-    add_index: bool = True,
-) -> list[dict[str, Any]]:
-    if extra_fields is None:
-        extra_fields = []
-
-    values = _get_values_for_table_repr(_self)
-    cols = defaultdict(list)
-    for item in values:
-        # unpack dict
-        if isinstance(_self, Mapping):
-            key, item = item
-            cols["key"].append(key)
-
-        # get id
-        id_ = getattr(item, "id", None)
-        include_id = getattr(item, "__syft_include_id_coll_repr__", True)
-        if id_ is not None and include_id:
-            cols["id"].append({"value": str(id_), "type": "clipboard"})
-
-        if type(item) == type:
-            t = full_name_with_qualname(item)
-        else:
-            try:
-                t = item.__class__.__name__
-            except Exception:
-                t = item.__repr__()
-
-        if not is_homogenous:
-            cols["type"].append(t)
-
-        # if has _coll_repr_
-
-        if hasattr(item, "_coll_repr_"):
-            ret_val = item._coll_repr_()
-            if "id" in ret_val:
-                del ret_val["id"]
-            for key in ret_val.keys():
-                cols[key].append(ret_val[key])
-        else:
-            for field in extra_fields:
-                value = item
-                try:
-                    attrs = field.split(".")
-                    for i, attr in enumerate(attrs):
-                        # find indexing like abc[1]
-                        res = re.search(r"\[[+-]?\d+\]", attr)
-                        has_index = False
-                        if res:
-                            has_index = True
-                            index_str = res.group()
-                            index = int(index_str.replace("[", "").replace("]", ""))
-                            attr = attr.replace(index_str, "")
-
-                        value = getattr(value, attr, None)
-                        if isinstance(value, list) and has_index:
-                            value = value[index]
-                        # If the object has a special representation when nested we will use that instead
-                        if (
-                            hasattr(value, "__repr_syft_nested__")
-                            and i == len(attrs) - 1
-                        ):
-                            value = value.__repr_syft_nested__()
-                        if (
-                            isinstance(value, list)
-                            and i == len(attrs) - 1
-                            and len(value) > 0
-                            and hasattr(value[0], "__repr_syft_nested__")
-                        ):
-                            value = [
-                                (
-                                    x.__repr_syft_nested__()
-                                    if hasattr(x, "__repr_syft_nested__")
-                                    else x
-                                )
-                                for x in value
-                            ]
-                    if value is None:
-                        value = "n/a"
-
-                except Exception as e:
-                    print(e)
-                    value = None
-                cols[field].append(str(value))
-
-    col_lengths = {len(cols[col]) for col in cols.keys()}
-    if len(col_lengths) != 1:
-        raise ValueError(
-            "Cannot create table for items with different number of fields."
-        )
-
-    num_rows = col_lengths.pop()
-    if add_index and TABLE_INDEX_KEY not in cols:
-        cols[TABLE_INDEX_KEY] = list(range(num_rows))
-
-    rows = []
-    for i in range(num_rows):
-        row = {}
-        for col in cols.keys():
-            row[col] = cols[col][i]
-        rows.append(row)
-
-    if is_homogenous:
-        sort_key = getattr(values[0], "__table_sort_attr__", None) or "created_at"
-    else:
-        sort_key = "created_at"
-    rows = _sort_table_rows(rows, sort_key)
-
-    return rows
-
-
-def _get_grid_template_columns(first_value: Any) -> tuple[str | None, str | None]:
-    grid_template_cols = getattr(first_value, "__table_coll_widths__", None)
-    if isinstance(grid_template_cols, list):
-        grid_template_columns = " ".join(grid_template_cols)
-        grid_template_cell_columns = "unset"
-    else:
-        grid_template_columns = None
-        grid_template_cell_columns = None
-    return grid_template_columns, grid_template_cell_columns
-
-
-def _prepare_table_data(
-    obj: Any,
-    add_index: bool = True,
-) -> tuple[list[dict], dict]:
-    """
-    Returns a tuple[list[dict], dict] of table_data, table_metadata
-    """
-
-    values = _get_values_for_table_repr(obj)
-    if len(values) == 0:
-        return [], {}
-
-    first_value = values[0]
-    if not _syft_in_mro(obj, first_value):
-        raise ValueError("Cannot create table for Non-syft objects.")
-
-    extra_fields = getattr(first_value, "__repr_attrs__", [])
-    is_homogenous = len({type(x) for x in values}) == 1
-    table_data = _get_table_data(
-        _self=obj,
-        is_homogenous=is_homogenous,
-        extra_fields=extra_fields,
-        add_index=add_index,
-    )
-
-    if is_homogenous:
-        cls_name = first_value.__class__.__name__
-        grid_template_columns, grid_template_cell_columns = _get_grid_template_columns(
-            first_value
-        )
-    else:
-        cls_name = ""
-        grid_template_columns = None
-        grid_template_cell_columns = None
-
-    table_icon = getattr(first_value, "icon", None)
-    table_name = f"{cls_name} {obj.__class__.__name__.capitalize()}"
-
-    table_metadata = {
-        "name": table_name,
-        "icon": table_icon,
-        "grid_template_columns": grid_template_columns,
-        "grid_template_cell_columns": grid_template_cell_columns,
-    }
-
-    return table_data, table_metadata
-
-
-def list_dict_repr_html(self: Mapping | Set | Iterable) -> str | None:
-    try:
-        table_data, table_metadata = _prepare_table_data(self)
-        if len(table_data) == 0:
-            # TODO # TODO cleanup tech debt: _repr_html_ is used in syft without `None` fallback.
-            return self.__repr__()
-        return create_table_template(
-            table_data=table_data,
-            **table_metadata,
-        )
-
-    except Exception as e:
-        logger.debug(f"Could not create table: {e}")
-
-    # _repr_html_ returns None -> fallback to default repr
-    return None
-
-
-def _get_values_for_table_repr(obj: Any) -> Any:
-    if isinstance(obj, Mapping):
-        values = list(obj.values())
-    elif isinstance(obj, Set):
-        values = list(obj)
-    else:
-        values = obj
-
-    return values
-
-
-def _syft_in_mro(self: Any, item: Any) -> bool:
-    if hasattr(type(item), "mro") and type(item) != type:
-        mro = type(item).mro()
-    elif hasattr(item, "mro") and type(item) != type:
-        mro = item.mro()
-    else:
-        mro = str(self)  # type: ignore
-
-    return "syft" in str(mro).lower()
 
 
 # give lists and dicts a _repr_html_ if they contain SyftObject's
